# server/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  handlers.DeleteWorkspaceResponse:
    properties:
      nextWorkspaceName:
        type: string
    type: object
  handlers.ErrorResponse:
    properties:
      message:
        type: string
    type: object
...
  handlers.LastWorkspaceNameResponse:
    properties:
      lastWorkspaceName:
        type: string
    type: object
...
  models.Workspace:
    properties:
      autoSave:
        type: boolean
      createdAt:
        type: string
      gitAutoCommit:
        type: boolean
      gitCommitEmail:
        type: string
      gitCommitMsgTemplate:
        type: string
      gitCommitName:
        type: string
      gitEnabled:
        type: boolean
      gitToken:
        type: string
      gitUrl:
        type: string
      gitUser:
        type: string
      id:
        minimum: 1
        type: integer
      lastOpenedFilePath:
        type: string
      name:
        type: string
      showHiddenFiles:
        type: boolean
      theme:
        description: Integrated settings
        enum:
        - light
        - dark
        type: string
      userId:
        minimum: 1
        type: integer
    required:
    - id
    - name
    - userId
    type: object
...
info:
  contact: {}
  description: This is the API for Lemma markdown note taking app.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Lemma API
  version: "1.0"
paths:
...
  /workspaces:
    get:
      description: Lists all workspaces for the current user
      operationId: listWorkspaces
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Workspace'
            type: array
        "500":
          description: Failed to list workspaces
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - CookieAuth: []
      summary: List workspaces
      tags:
      - workspaces
    post:
      consumes:
      - application/json
      description: Creates a new workspace
      operationId: createWorkspace
      parameters:
      - description: Workspace
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.Workspace'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Workspace'
        "400":
          description: Invalid workspace
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Failed to setup git repo
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - CookieAuth: []
      summary: Create workspace
      tags:
      - workspaces
  /workspaces/{workspace_name}:
    delete:
      description: Deletes the current workspace
      operationId: deleteWorkspace
      parameters:
      - description: Workspace name
        in: path
        name: workspace_name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.DeleteWorkspaceResponse'
        "400":
          description: Cannot delete the last workspace
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Failed to commit transaction
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - CookieAuth: []
      summary: Delete workspace
      tags:
      - workspaces
    get:
      description: Returns the current workspace
      operationId: getWorkspace
      parameters:
      - description: Workspace name
        in: path
        name: workspace_name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Workspace'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - CookieAuth: []
      summary: Get workspace
      tags:
      - workspaces
    put:
      consumes:
      - application/json
      description: Updates the current workspace
      operationId: updateWorkspace
      parameters:
      - description: Workspace name
        in: path
        name: workspace_name
        required: true
        type: string
      - description: Workspace
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.Workspace'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Workspace'
        "400":
          description: Invalid request body
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Failed to setup git repo
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - CookieAuth: []
      summary: Update workspace
      tags:
      - workspaces
...
  /workspaces/last:
    get:
      description: Returns the name of the last opened workspace
      operationId: getLastWorkspaceName
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.LastWorkspaceNameResponse'
        "500":
          description: Failed to get last workspace
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - CookieAuth: []
      summary: Get last workspace name
      tags:
      - workspaces
    put:
      consumes:
      - application/json
      description: Updates the name of the last opened workspace
      operationId: updateLastWorkspaceName
      produces:
      - application/json
      responses:
        "204":
          description: No Content - Last workspace updated successfully
        "400":
          description: Invalid request body
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Failed to update last workspace
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - CookieAuth: []
      summary: Update last workspace name
      tags:
      - workspaces
securityDefinitions:
  CookieAuth:
    in: cookie
    name: access_token
    type: apiKey
swagger: "2.0"